Added back devices and ports info in config.md
Added back the devices and ports configuration info in the config file. I came to the github repo looking for this exact information since I changed my USB headphones and wanted to reconfigure but I'm still running an older version. I don't think it clatters the documentation too much and it might still be needed.